chore(seeding): backend seeding restructure implemented
