+++ apps/backend/src/db/database-init.service.ts
@@ -1,12 +1,15 @@
 import { Injectable, Logger, OnApplicationBootstrap } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
+// import { ConfigService } from '@nestjs/config';
 import { sql } from 'drizzle-orm';
 import { DatabaseService } from './database.service';
-import { DatabaseSeedService } from './database-seed.service';
+// import { DatabaseSeedService } from './seeding/database-seed.service';
 import { migrate as migrateNodePg } from 'drizzle-orm/node-postgres/migrator';
 import { migrate as migratePglite } from 'drizzle-orm/pglite/migrator';
 import { join } from 'node:path';
-import { DB_MODES, parseSeedFlag } from './database.constants';
+import {
+  DB_MODES,
+  //  parseSeedFlag
+} from './database.constants';
 import type { NodePgDatabase } from 'drizzle-orm/node-postgres';
 import type { PgliteDatabase } from 'drizzle-orm/pglite';
 
@@ -16,13 +19,13 @@ export class DatabaseInitService implements OnApplicationBootstrap {
 
   constructor(
     private readonly databaseService: DatabaseService,
-    private readonly configService: ConfigService,
-    private readonly databaseSeedService: DatabaseSeedService,
+    // private readonly configService: ConfigService,
+    // private readonly databaseSeedService: DatabaseSeedService,
   ) {}
 
   async onApplicationBootstrap(): Promise<void> {
     const dbMode = this.databaseService.dbMode;
-    const shouldSeed = parseSeedFlag(this.configService.get<string>('SEED'));
+    // const shouldSeed = parseSeedFlag(this.configService.get<string>('SEED'));
     const migrationsFolder = join(process.cwd(), 'drizzle');
 
     try {
@@ -51,9 +54,9 @@ export class DatabaseInitService implements OnApplicationBootstrap {
       }
       this.logger.log('Database migrations applied successfully');
 
-      if (shouldSeed) {
-        await this.databaseSeedService.seed();
-      }
+      // if (shouldSeed) {
+      //   await this.databaseSeedService.seed();
+      // }
     } catch (error) {
       this.logger.error('Failed to run database migrations', error);
       throw error;
