+++ apps/backend/src/db/seeding/Constants/Courses.constants.ts
