+++ apps/backend/src/db/seeding/services/index.ts
@@ -1,0 +1,5 @@
+export * from './university.seed.service';
+export * from './users.seed.service';
+export * from './universityRoles.seed.service';
+export * from './courses.seed.service';
+export * from './modules.seed.service';
