--- apps/backend/src/db/seeds/README.md
@@ -1,6 +1,0 @@
-/\*\*
-
-- Seeds folder for database seeding migrations
-- Each seed file implements SeedMigration interface
-- Register new seeds in database-seed.service.ts
-  \*/
