--- apps/backend/src/db/seeds/index.ts
@@ -1,1 +1,0 @@
-export * from './Constants';
